basic recipe adding/viewing functionality added
A user that is logged in can add a recipe to the database.
Any user, be them logged in or a guest, can view the list of recipes by
clicking on "recipes" in the navigation bar. Clicking on a recipe brings
you to that recipe, and basic information about that recipe is shown.

IMPORTANT: You must redo all the migrations and seed your database again
for this commit to work.